+++ src/outbox/index.ts
@@ -1,6 +1,16 @@
 export type { DomainEventPublisher } from './DomainEventPublisher.js';
 export { EventBusDomainEventPublisher } from './DomainEventPublisher.js';
 export { InMemoryOutboxRepository } from './InMemoryOutboxRepository.js';
+export type {
+	DomainEventConstructor,
+	OutboxEventRegistration,
+	RegisteredOutboxEvent,
+} from './OutboxEventRegistry.js';
+export { OutboxEventRegistry } from './OutboxEventRegistry.js';
+export {
+	OutboxEventAlreadyRegisteredError,
+	OutboxEventNotRegisteredError,
+} from './OutboxEventRegistryErrors.js';
 export * from './OutboxMessage.js';
 export type {
 	Clock,
@@ -22,3 +32,5 @@ export type {
 	FindPendingOutboxMessagesParams,
 	OutboxRepository,
 } from './OutboxRepository.js';
+export { RegisteredOutboxMessageDeserializer } from './RegisteredOutboxMessageDeserializer.js';
+export { RegisteredOutboxMessageSerializer } from './RegisteredOutboxMessageSerializer.js';